added dynamic exercise routing and restructured api folders and edited query function to allow array of strings
